// generated_new/user/model/login-user-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface LoginUserVO {
    'id'?: string;
    'userName'?: string;
    'userAvatar'?: string;
    'userProfile'?: string;
    'userRole'?: string;
    'phone'?: string;
    'createTime'?: string;
    'updateTime'?: string;
}